Add emote selection for 8ball
!8ball will now randomly add emotes to the end of responses. This is
currently set to a 90% chance, with a different set of emotes available
for each response type. To help contain specialized responses such as
"N OMEGALUL ", response objects now have a suffix property determining
whether a suffix can be affixed.

Due to the need for another property, weighted_pick was changed to
return the object instead of the response within, and string formatting
is now done within eightball